requirements.txt
asgiref==3.8.1
Django==5.0.6
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.13.2
django-admin-sortable2==2.2.3
django-ckeditor==6.7.1
django-filter==24.3
django-jazzmin==3.0.0
django-js-asset==2.2.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
